improved monster behavior, added 2 types of monsters, basic testing stage avaliable, started working on moving everything to room(undo hardcode) bullets WIP

+++ src/main/java/edu/pacific/comp55/starter/Point.java
@@ -1,29 +1,29 @@
 package edu.pacific.comp55.starter;
 
 public class Point {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
 	
 	
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		super();
 		this.x = x;
 		this.y = y;
 	}
-	public int getX() {
+	public double getX() {
 		return x;
 	}
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
-	public void setXY(int x, int y) {
+	public void setXY(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
